requirements.txt
# requirements.txt

# Document Processing
pdfminer.six==20221105
python-docx==0.8.11
chardet==5.1.0

# NLP and Text Processing
nltk==3.9.1
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3

# Apache Spark
pyspark==3.4.1

# Embeddings
sentence-transformers==3.3.1
torch==2.9.0
transformers==4.46.3

# Vector Databases
chromadb==0.5.23
faiss-cpu==1.9.0.post1  # or faiss-gpu for GPU support

# Web Framework
flask==3.1.0
flask-cors==5.0.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0